updated project

+++ src/main/java/com/coffeeshop/menu/CoffeeShopMenuApplication.java
@@ -3,6 +3,7 @@ package com.coffeeshop.menu;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ApplicationContext;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 import com.coffeeshop.menu.model.Product;
 
@@ -12,8 +13,8 @@ public class CoffeeShopMenuApplication {
         ApplicationContext context = SpringApplication.run(CoffeeShopMenuApplication.class, args);
 
         // Product instantiation and printing details
-        Product espresso = new Product(1, "Espresso", 2.99);
-        Product croissant = new Product(2, "Croissant", 3.49);
+        Product espresso = new Product((long)1, "Espresso", 2.99);
+        Product croissant = new Product((long)2, "Croissant", 3.49);
     
         System.out.println("Product ID: " + espresso.getId() + " Product Name: " + espresso.getName() + " Product Price: " + espresso.getPrice());
         System.out.println("Product ID: " + croissant.getId() + " Product Name: " + croissant.getName() + " Product Price: " + croissant.getPrice());
